fix: correct formula in fahrenheit_to_celsius function for accurate conversion

+++ temperature_conv/src/lib.rs
@@ -1,7 +1,7 @@
 pub fn fahrenheit_to_celsius(f: f64) -> f64 {
-    (f - 32 as f64) / (9 as f64/ 5 as f64)
+    (f -32.0) / (9.0 / 5.0)
 }
 
 pub fn celsius_to_fahrenheit(c: f64) -> f64 {
-    c * (9 as f64 / 5 as f64) + 32 as f64
+    (c * (9 as f64 / 5 as f64)) + 32 as f64
 }
